pyproject.toml
[tool.poetry]
name = "farmers-edge-server"
version = "0.1.0"
description = ""
authors = ["manulangat <emmanuelthedeveloper@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.6"
django-rest-passwordreset = "^1.1.0"
django-cors-headers = "^3.3.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
python-decouple = "^3.3"
six = "^1.14.0"
django-rest-framework = "^0.1.0"
django-rest-knox = "^4.1.0"
Celery = "^4.4.2"
pytest-django = "^3.9.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
